requirements.txt
# Core ML/RL libraries
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
ray[rllib]>=2.8.0

# Visualization and UI
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.17.0

# Configuration and utilities
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.16.0
tqdm>=4.66.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0
